+++ dflib/src/test/java/org/dflib/RowSet_Merge_ExpTest.java
@@ -70,12 +70,11 @@ public class RowSet_Merge_ExpTest {
 
     @Test
     public void byIndex_Duplicate() {
-        DataFrame df = DataFrame.foldByRow("a", "b", "c")
+        DataFrame df = DataFrame.byColumn("a", "b", "c")
                 .of(
-                        1, "x", "a",
-                        2, "y", "b",
-                        -1, "m", "n")
-                .cols(0).compactInt(0)
+                        Series.ofInt(1, 2, -1),
+                        Series.of("x", "y", "m"),
+                        Series.of("a", "b", "n"))
                 .rows(Series.ofInt(0, 2, 2, 0))
                 .merge(
                         Exp.$int(0).mul(3),